First phase of upgrade to parallel CR 3.0.2.
Merge completed, conflicts resolved, builds on 32-bit Linux in
src/main, but not yet further tested.


git-svn-id: http://projects.cs.kent.ac.uk/projects/cxxr/svn/trunk@1461 549f2c98-2231-0410-b618-a09cd5e121c0

+++ src/modules/internet/sockconn.cpp
@@ -16,7 +16,7 @@
 
 /*
  *  R : A Computer Language for Statistical Data Analysis
- *  Copyright (C)  2001-5   The R Core Team.
+ *  Copyright (C)  2001-12   The R Core Team.
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -95,7 +95,7 @@ static Rboolean sock_open(Rconnection con)
     }
     thisconn->fd = sock;
 
-    mlen = strlen(con->mode);
+    mlen = int( strlen(con->mode));
     con->isopen = TRUE;
     if(mlen >= 2 && con->mode[mlen - 1] == 'b') con->text = FALSE;
     else con->text = TRUE;
@@ -111,11 +111,11 @@ static void sock_close(Rconnection con)
     con->isopen = FALSE;
 }
 
-static int sock_read_helper(Rconnection con, void *ptr, size_t size)
+static size_t sock_read_helper(Rconnection con, void *ptr, size_t size)
 {
     Rsockconn thisconn = (Rsockconn)con->connprivate;
-    int res;
-    int nread = 0, n;
+    ssize_t res;
+    size_t nread = 0, n;
 
     con->incomplete = FALSE;
     do {
@@ -155,7 +155,7 @@ static int sock_read_helper(Rconnection con, void *ptr, size_t size)
 static int sock_fgetc_internal(Rconnection con)
 {
     unsigned char c;
-    int n;
+    size_t n;
 
     n = sock_read_helper(con, (char *)&c, 1);
     return (n == 1) ? c : R_EOF;
